+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Widgets can now be hidden through a new configuration page ([#40](https://github.com/scm-manager/scm-landingpage-plugin/pull/40))
+
+### Changed
+- Integrate plugin content into repository overview page ([#40](https://github.com/scm-manager/scm-landingpage-plugin/pull/40))
+
 ### 1.6.0 - 2021-07-30
 ### Changed
 - Fix favorite repository layout to match new repository entry ([#35](https://github.com/scm-manager/scm-landingpage-plugin/pull/35))